requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-jose[cryptography]
python-multipart
requests